% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{arcene}
\alias{arcene}
\title{arcene}
\format{An object of class \code{list} of length 2.}
\source{
https://statweb.stanford.edu/~tibs/strong/

https://archive.ics.uci.edu/ml/datasets/Arcene
}
\usage{
arcene
}
\description{
Binary response data set.
}
\section{UCI abstract}{
 ARCENE's task is to distinguish cancer
versus normal patterns from
mass-spectrometric data. This is a two-class classification problem
with continuous input variables. This dataset is one of 5 datasets
of the NIPS 2003 feature selection challenge.
}

\keyword{datasets}
