% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/buildDataAbstraction.R
\name{buildDataAbstraction}
\alias{buildDataAbstraction}
\title{Dataset abstraction builder}
\usage{
buildDataAbstraction(name, data, labels)
}
\arguments{
\item{name}{a one element character vector containing the data object name.}

\item{data}{a matrix or data.frame containing the attributes data.}

\item{labels}{a vector or factor representing the labels of the dataset.}
}
\description{
Helper function to generate an abstraction for a
specific dataset, according to the interface
pattern used in \code{\link{rDatasets}} package.
}
\examples{
data(iris)

iris_ = buildDataAbstraction(name = "Iris", data = iris[, -5], labels = iris[, 5])
}

