% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cpusmall}
\alias{cpusmall}
\title{cpusmall}
\format{
An object of class \code{list} of length 2.
}
\source{
http://www.cs.toronto.edu/~delve/data/datasets.html

https://www.csie.ntu.edu.tw/~cjlin/libsvmtools/datasets/regression.html
}
\usage{
cpusmall
}
\description{
Regression data set from the LIBSVM database
}
\keyword{datasets}
