% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dorothea}
\alias{dorothea}
\title{Dorothea}
\format{An object of class \code{list} of length 2.}
\source{
https://archive.ics.uci.edu/ml/datasets/Dorothea

https://statweb.stanford.edu/~tibs/strong/
}
\usage{
dorothea
}
\description{
Binary response data set.
}
\section{UCI abstract}{
 DOROTHEA is a drug discovery dataset. Chemical
compounds represented
by structural molecular features must be classified as active (binding
to thrombin) or inactive. This is one of 5 datasets of the NIPS 2003
feature selection challenge.
}

\keyword{datasets}
