% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{e2006_tfidf}
\alias{e2006_tfidf}
\title{E2006-tfidf}
\format{An object of class \code{list} of length 2.}
\source{
http://www.ark.cs.cmu.edu/10K/

https://www.csie.ntu.edu.tw/~cjlin/libsvmtools/datasets/regression.html
}
\usage{
e2006_tfidf
}
\description{
Regression data set from LIBSVM. Only the test data set
is used here. From LIBSVM: "Features include the volitility in the past 12
months and tf-idf of unigrams."
}
\keyword{datasets}
