% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gisette}
\alias{gisette}
\title{Gisette}
\format{An object of class \code{list} of length 2.}
\source{
https://statweb.stanford.edu/~tibs/strong/

https://archive.ics.uci.edu/ml/datasets/Gisette
}
\usage{
gisette
}
\description{
Gisette
}
\section{UCI abstract}{
 GISETTE is a handwritten digit recognition problem.
The problem is to separate the highly confusible digits '4' and '9'. This
dataset is one of five datasets of the NIPS 2003 feature selection challenge.
}

\keyword{datasets}
