% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{golub}
\alias{golub}
\title{Golub}
\format{An object of class \code{list} of length 2.}
\source{
https://statweb.stanford.edu/~tibs/strong/
}
\usage{
golub
}
\description{
Gene expression data set. Binary response.
}
\keyword{datasets}
