% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{news20}
\alias{news20}
\title{News20}
\format{An object of class \code{list} of length 2.}
\source{
Ken Lang. Newsweeder: Learning to filter netnews.
In Proceedings of the Twelfth International Conference on Machine
Learning, pages 331-339, 1995.

https://www.csie.ntu.edu.tw/~cjlin/libsvmtools/datasets/multiclass.html
}
\usage{
news20
}
\description{
Multi-Class training data set from the LIBSVM database
}
\keyword{datasets}
