% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdatasets-package.R
\docType{package}
\name{rdatasets-package}
\alias{rdatasets}
\alias{rdatasets-package}
\title{rdatasets: Datasets for R}
\description{
Several datasets formatted for straightforward use in
    testing out statistical models, particularly when it comes to benchmarking.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jolars/rdatasets}
  \item Report bugs at \url{https://github.com/jolars/rdatasets/issues}
}

}
\author{
\strong{Maintainer}: Johan Larsson \email{johanlarsson@outlook.com} (\href{https://orcid.org/0000-0002-4029-5945}{ORCID})

}
\keyword{internal}
