% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zipcode}
\alias{zipcode}
\title{zipcode}
\format{An object of class \code{list} of length 2.}
\source{
https://web.stanford.edu/~hastie/ElemStatLearn/

https://www.csie.ntu.edu.tw/~cjlin/libsvmtools/datasets/regression.html
}
\usage{
zipcode
}
\description{
Multi-class training data set.
}
\section{From source}{
 Normalized handwritten digits, automatically
scanned from envelopes by the U.S. Postal Service. The original
scanned digits are binary and of different sizes and orientations; the
images  here have been deslanted and size normalized, resulting
in 16 x 16 grayscale images (Le Cun et al., 1990).
}

\keyword{datasets}
