% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{physician}
\alias{physician}
\title{physician}
\format{An object of class \code{list} of length 2.}
\source{
Deb, Partha, and Pravin K. Trivedi. "Demand for Medical Care by
the Elderly: A Finite Mixture Approach." Journal of Applied Econometrics,
vol. 12, no. 3, 1997, pp. 313–336. JSTOR, www.jstor.org/stable/2285252.

https://www.jstatsoft.org/article/view/v027i08
}
\usage{
physician
}
\description{
Count outcome data set related to a paper by Deb et al. The variable
\code{ofp} (physician office visits) have been used as the outcome here.
}
\keyword{datasets}
